requirements.txt
discord.py
cassiopeia
html2text
fuzzywuzzy
